backend/requirements.txt
django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
pyjwt==1.7.1
